New model API.
The first parameter of model.task is now a Task_spec and inputs is its
attribute.  This should be more extensible: if we need to pass more stuff
to the task we add more attributes.  No need to modify model.task unless
it needs to care about the new stuff.