Test and fix Socket.IO.
Looks like emit only works in the server thread when using eventlet.
So put in a little queue also for emits and a coroutine to actually
emit them in the server thread.  Unfortunately waiting for the
queue depends on async_mode, thus only eventlet and threading (where
no queue appears needed) are supported for now.