Use client.gather in TaskFlask.flush_updates.
The main task callback now adds the future to the dict
TaskFlask.gathers.  flush_updates gathers the whole dict with
client.gather, which may be more efficient than fetching results
one at a time.  It then clears the dict and processes the queue.
If tasks complete at the same time, it is possible that their results
get fetched outside the gather call, then their future is added to
gathers although it has already been saved.  That may cause some
inefficiency but should not cause errors.

Updates can now be tuples: the first member is called, the others are
parameters (in addition to conn).  This is not yet used though; it
may need further improvement to be useful.