An optimisation test problem.
This is just a small (n = 20) mean-variance portfolio optimization
problem.  The objectives (portfolio mean and variance) are computed
directly in Python, which is ridiculously fast.  Thus it really puts
the I/O through its paces.  Had to upgrade dask.distributed to 1.24.1
and Tornado to 5.1.1 before this started to work properly.