Simsvc Docker image.
The Makefile builds a wheel, then installs it in a Docker image.
Linux only for now: we use daskdev/dask as the base image.

run-server.py was moved to simsvc/__main__.py and can now
be executed with python3 -m simsvc.  This avoids full pathnames:
the base image has Python at an unusual location.