Now runs on Azure.
AKS provides preconfigured storage classes only for Azure Disks, which
do not support RWX access.  For that we need Azure Files, which requires
more config.  An Azure storage account must be created (here we called
it intenswork) and a Kubernetes storage class that uses it must be
defined.  Even some RBAC config is needed.  The label selector doesn't
appear to work with the new storage class, so drop that from the chart.