More client tweaking.
Get rid of error_jobs.  On error put the job back in jobs instead,
so it continues to be waited for and periodically polled.  Maybe
it was a temporary error and we'll succeed getting results eventually.

If a job is missing from the server, try again later.  This seems
to happen sometimes if the job is queried very soon after posting.
No idea why - the server code looks tight to me.

Stop treating InterruptedIOException like InterruptedException.
Treat it like any other IOException instead.  The most common IIOExc
is a socket timeout.

Wait for Socket.IO connection in start rather than the constructor.