Put the job database on a PV.
Using the same storage class as for the work volume because
could not get Azure disks to work.  The problem is probably that
the disk has an actual Unix filesystem with permission bits and
the Simsvc user (nobody:nogroup) has no write access.  That cannot
be fixed at build time.