First successful Virtual Kubelet test.
Moved to West Europe.  Simplified the worker template to include
a worker.spec value, which replaced tolerations and nodeSelector.
It can also include any other pod spec fields that are not in the
template - such as a dnsConfig to patch issue #456 of VK.

The test was with a rather manually set up cluster, roughly according
to aks/vk-deploy.sh.