HTTP basic authentication.
Authentication is enabled by setting env. var. HTPASSWD_FILE
to a htpasswd file.  It is better to handle authentication at the
ingress server, which also terminates SSL.  Unfortunately that is
not possible with OpenShift routes and we need to do it in the back
end.  Passwords will then be sent unencrypted on the cluster private
network (but not elsewhere if secure routes are used).

Requires passlib and flask_httpauth.  Bump Simsvc version to 0.3.